requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.0+cu121
torchvision==0.19.0+cu121
torchaudio==2.4.0+cu121
matplotlib==3.7.2
numpy==1.24.3
opencv-python==4.9.0.80
Pillow
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git